requirements.txt
ipykernel
ipywidgets
python-dotenv
numpy
pillow
tqdm
google-generativeai
langchain-google-genai
langchain
langchain-community
chromadb
pypdf
pandas
fastapi[standard]